requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-timezone-field==7.1
docker==7.1.0
greenlet==3.1.1
hiredis==3.1.0
idna==3.10
kombu==5.4.2
numpy==2.2.2
pandas==2.2.3
prompt_toolkit==3.0.50
psycopg2==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
redis==5.2.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
